Duck-type VQE interim results (#257)
* Duck-type VQE interim results

Add a check for the type of data expected in the callback wraper in VQEProgram.

This will allow us to close Qiskit-Partners/qiskit-runtime#38

Anticipating changes in the Qiskit Runtime API which will start emitting final results in the WebSocket stream. When that happens, if we don't duck-type in the VQE callback wrapper receiving the interim results, it will get a ValueError exception.

Already done in Qiskit/qiskit-nature#350

* Update qiskit_optimization/runtime/vqe_program.py

fix typo in doc string

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>
Co-authored-by: Takashi Imamichi <31178928+t-imamichi@users.noreply.github.com>